+++ src/main/java/com/comeon/backend/user/domain/SocialLoginHandler.java
@@ -2,16 +2,14 @@ package com.comeon.backend.user.domain;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
-@Transactional
 @RequiredArgsConstructor
-public class UserService {
+public class SocialLoginHandler {
 
     private final UserRepository userRepository;
 
-    public User saveUser(String oauthId, OauthProvider provider, String email, String name) {
+    public User doLogin(String oauthId, OauthProvider provider, String email, String name) {
         User user = userRepository.findByOauthIdAndProvider(oauthId, provider)
                 .orElse(
                         User.builder()
